requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2-binary==2.8.5
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
